requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-phonenumber-field==5.0.0
phonenumbers==8.12.11
pytz==2020.1
sqlparse==0.4.1
